ai-support/python/requirements.txt
# Rendly AI Support - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# NLP / Text Processing
rapidfuzz>=3.6.1

# Machine Learning
scikit-learn>=1.4.0
numpy>=1.26.0

# Caching
cachetools>=5.3.2

# HTTP Client (for LLM API calls)
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0
